Fix sticking of stacked objects
Because matplotlib.lines.Line2D.contains() does not account for render order, if draggable lines are close enough together the event callbacks will fire for both simultaneously, causing the lines to stick together. To mitigate this, a list of objects whose motion callbacks will fire is generated and priority is given to the last object created, assumed to be the topmost rendered object.