Add indicator for internal classes, fix disconnect
Add weak indicators to all internal classes and update internal calls
appropriately. This slightly mitigates confusion on which class methods
to call in order to get draggable objects on the screen.

Fix instaces of disconnect where () was not present so it wasn't
actually being called to disconnect the motion callbacks

Fix DragWedge's patch call to use center instead of xy, which isn't an
input to the method