Change position setting from EAFP to LBYL
Utilizing 'Easier to Ask For Permission' while attempting to update an
object location doesn't work for this use case because asking will
create the attribute if it doesn't exist and an AttributeError will
never be raised. This results in patches without a 'center' attribute,
like rectangles, not exhibiting the desired behavior (dragging, that's
an important one). 'Look Before You Leap' is used instead, verifying the
attribute exists before trying to set it. This restores behavior to the
affected patch types.