Fix Wedge position update
For *reasons*, matplotlib.patches.Wedge does not shift position when its
'center' attribute is modified. It does change positions when the update
method is called, so this has been added in order to make the drag
wedges actually draggable. update throws AttributeErrors for (most)
other patches so the existing LBYL has been retained as a catch for
update not working.

Resolves: #1